Replace axios with custom api
