+++ frontend/pages/user/index.tsx
@@ -1,6 +1,8 @@
+import UserCard from '@/components/user/userCard';
+
 //유저 페이지 입니다. 경로 '/user/'
 const User = () => {
-  return <div>유저 페이지</div>;
+  return <UserCard />;
 };
 
 export default User;
